chapter_001/src/main/java/ru/job4j/collections/ReversedLinked.java
package ru.job4j.collections;

import java.util.Iterator;
import java.util.NoSuchElementException;

public class ReversedLinked<T> implements Iterable<T> {
    private Node<T> head;

    public void add(T value) {
        Node<T> node = new Node<T>(value, null);
        if (head == null) {
            head = node;
            return;
        }
        Node<T> tail = head;
        while (tail.next != null) {
            tail = tail.next;
        }
        tail.next = node;
    }

    public void revert() {
        Node<T> temp = head;
        Node<T> previous = null;
        Node<T> current = null;
        while (temp != null) {
            current = temp;
            temp = temp.next;
            current.next = previous;
            previous = current;
            head = current;
        }
    }

    @Override
    public Iterator<T> iterator() {
        return new Iterator<T>() {
            private Node<T> node = head;

            @Override
            public boolean hasNext() {
                return node != null;
            }

            @Override
            public T next() {
                if (!hasNext()) {
                    throw new NoSuchElementException();
                }
                T value = node.value;
                node = node.next;
                return value;
            }
        };
    }

    private static class Node<T> {
       private T value;
       private Node<T> next;

        public Node(T value, Node<T> next) {
            this.value = value;
            this.next = next;
        }
    }
}
